requirements.txt
Django==2.1.4
py-cpuinfo==4.0.0
scipy==1.0.1
six==1.11.0